+++ app/controllers/notifications/application_controller.rb
@@ -3,7 +3,9 @@ module Notifications
     helper_method :current_user
 
     alias_method :origin_current_user, Notifications.config.current_user_method.to_sym
-    alias_method :origin_authenticate_user!, Notifications.config.authenticate_user_method.to_sym
+    if Notifications.config.authenticate_user_method
+      alias_method :origin_authenticate_user!, Notifications.config.authenticate_user_method.to_sym
+    end
 
     before_action :authenticate_user!
 
@@ -12,7 +14,7 @@ module Notifications
     end
 
     def authenticate_user!
-      origin_authenticate_user!
+      origin_authenticate_user! if Notifications.config.authenticate_user_method
     end
   end
 end