+++ test/test_helper.rb
@@ -25,9 +25,9 @@ require 'minitest/mock'
 
 FactoryGirl.find_definitions
 
-# Homeland.configure do
-#   self.user_avatar_url_method = 'avatar_url'
-# end
+Notifications.configure do
+  self.authenticate_user_method = 'authenticate_user!'
+end
 
 # Filter out Minitest backtrace while allowing backtrace from other libraries
 # to be shown.